include specific helpers instead of database::#{type}
This is because community cookbook database removed
database::mysql in v4.0.2, and database::postgresql
is a mere include of postgresql::ruby anyways.

In case of mysql, we now include mysql2_chef_gem
as directed by the database cookbook.
For postsgresql we just include postgresql::ruby
ourselves, since that may be removed next…